wireless: fix rtw_hal_ps_pwr_lvl_cfg req_pwr_lvl type (u32 -> u8)
The hal_api.h prototype declared req_pwr_lvl as u32, but the definition
and every consumer treat it as u8:

  definition rtw_hal_ps_pwr_lvl_cfg() : u8 req_pwr_lvl  (hal_ps.c)
  callee _hal_ps_pwr_lvl_cfg()        : u8 req_pwr_lvl
  callers / helpers in phl_ps.c       : u8 cur_pwr_lvl, u8 req_pwr_lvl
  phl_ps_pwr_lvl_to_str(), _judge_*() : u8

Power level is a small enumerated index, never wider than u8; the u32 in
the prototype was a typo. Narrow the prototype to u8 to match the
definition and all users.

+++ phl/hal_g6/hal_api.h
@@ -993,7 +993,7 @@ void rtw_hal_final_cap_decision(struct rtw_phl_com_t *phl_com, void *hal);
 /* Power Save APIs */
 enum rtw_hal_status
 rtw_hal_ps_pwr_lvl_cfg(struct rtw_phl_com_t *phl_com, void *hal,
-			u32 req_pwr_lvl);
+			u8 req_pwr_lvl);
 enum rtw_hal_status
 rtw_hal_ps_lps_cfg(void *hal, struct rtw_hal_lps_info *lps_info);
 enum rtw_hal_status
